refactor: put each request handler in its own file
issue: #14
