+++ include/silo_api/query_handler.h
@@ -4,7 +4,7 @@
 #include <Poco/Net/HTTPServerRequest.h>
 #include <Poco/Net/HTTPServerResponse.h>
 
-#include "silo_api/request_handler.h"
+#include "silo_api/rest_resource.h"
 
 namespace silo {
 class QueryEngine;
