+++ src/silo_api/query_handler.cpp
@@ -11,7 +11,7 @@
 #include "silo/query_engine/query_engine.h"
 #include "silo/query_engine/query_parse_exception.h"
 #include "silo/query_engine/query_result.h"
-#include "silo_api/request_handler.h"
+#include "silo_api/error_request_handler.h"
 #include "silo_api/variant_json_serializer.h"
 
 namespace silo::response {
